tools/tfrecord/requirements.txt
tensorflow-cpu
vlkit
pillow
torch
protobuf
